New sample file with database connection
This is a basic plug-and-play file that merchants can use to quickly set up a notification server. This will help them to save the push notifications from Adyen into a MySQL database. All merchants will need to do is setup a MySQL server and database. The code will take care of the table creation. The contents can then be pulled with MySQL requests.